Organize std and portable endpoint modules
