+++ crates/minip2p/src/std/discovery.rs
@@ -1,4 +1,4 @@
-//! Endpoint coordination for discovery sources and NAT traversal.
+//! Std endpoint coordination for discovery sources and NAT traversal.
 
 use std::collections::{BTreeMap, VecDeque};
 use std::time::Instant;
